+++ src/stats.rs
@@ -34,6 +34,16 @@ pub enum Element {
     DateTime(NaiveDateTime),
 }
 
+#[derive(Debug, PartialEq, Clone, Serialize, Deserialize, Eq, Hash)]
+pub enum GroupElement {
+    Int(i64),
+    UInt(u32),
+    Enum(String),
+    Text(String),
+    IpAddr(IpAddr),
+    DateTime(NaiveDateTime),
+}
+
 #[derive(Debug, Default, PartialEq, Clone, Serialize, Deserialize)]
 pub struct FloatRange {
     pub smallest: f64,
@@ -57,6 +67,19 @@ impl fmt::Display for Element {
             }
             Self::IpAddr(x) => write!(f, "{}", x),
             Self::DateTime(x) => write!(f, "{}", x),
+        }
+    }
+}
+
+impl PartialOrd for GroupElement {
+    fn partial_cmp(&self, other: &GroupElement) -> Option<std::cmp::Ordering> {
+        match (self, other) {
+            (Self::Int(s), Self::Int(o)) => Some(s.cmp(o)),
+            (Self::UInt(s), Self::UInt(o)) => Some(s.cmp(o)),
+            (Self::Enum(s), Self::Enum(o)) | (Self::Text(s), Self::Text(o)) => Some(s.cmp(o)),
+            (Self::IpAddr(s), Self::IpAddr(o)) => Some(s.cmp(o)),
+            (Self::DateTime(s), Self::DateTime(o)) => Some(s.cmp(o)),
+            _ => None,
         }
     }
 }
@@ -88,6 +111,18 @@ pub struct NLargestCount {
     pub(crate) number_of_elements: usize,
     pub(crate) top_n: Option<Vec<ElementCount>>,
     pub(crate) mode: Option<Element>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct GroupElementCount {
+    pub value: GroupElement,
+    pub count: usize,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct GroupCount {
+    pub count_index: Option<usize>, // if None, count just rows. If Some, count values of the column.
+    pub series: Vec<GroupElementCount>,
 }
 
 impl fmt::Display for Description {
@@ -487,7 +522,26 @@ pub(crate) fn n_largest_count_datetime(
     number_of_top_n: u32,
 ) -> NLargestCount {
     let mut n_largest_count = NLargestCount::default();
+    let values = convert_time_intervals(column, rows, time_interval);
 
+    top_n!(
+        values.iter(),
+        rows.len(),
+        n_largest_count,
+        NaiveDateTime,
+        Element::DateTime,
+        number_of_top_n
+    );
+
+    n_largest_count
+}
+
+#[must_use]
+pub(crate) fn convert_time_intervals(
+    column: &Column,
+    rows: &[usize],
+    time_interval: u32,
+) -> Vec<NaiveDateTime> {
     let time_interval = if time_interval > MAX_TIME_INTERVAL {
         MAX_TIME_INTERVAL
     // Users want to see time series of the same order intervals within MAX_TIME_INTERVAL which is in date units.
@@ -502,7 +556,7 @@ pub(crate) fn n_largest_count_datetime(
     };
     let time_interval = i64::from(time_interval);
 
-    let values = column
+    column
         .view_iter::<Int64ArrayType, i64>(rows)
         .unwrap()
         .map(|v: &i64| {
@@ -511,18 +565,7 @@ pub(crate) fn n_largest_count_datetime(
             ts += (*v - ts) / time_interval * time_interval;
             NaiveDateTime::from_timestamp(ts, 0)
         })
-        .collect::<Vec<_>>();
-
-    top_n!(
-        values.iter(),
-        rows.len(),
-        n_largest_count,
-        NaiveDateTime,
-        Element::DateTime,
-        number_of_top_n
-    );
-
-    n_largest_count
+        .collect::<Vec<_>>()
 }
 
 fn count_sort<I>(iter: I) -> Vec<(I::Item, usize)>